Add flag that abbreviates the size of the body to the given length
An issue that comes up when serving binary content is that the
response object will get logged and print out the entire image/
buffer/what have you as a utf-encoded string. This tends to A)
fill up the entire terminal which makes it hard to read the rest
of your logged messages, and B) add the cost of printing a 10k
char long string to the console to the runtime of your lambda
function.